0x11. C - printf README commit

+++ _printf_num.c
@@ -8,8 +8,8 @@
  */
 unsigned int printf_di(va_list val, int counter)
 {
-	int long num = va_arg(val, long int);
-	unsigned long int revNum;
+	int num = va_arg(val, int);
+	int revNum;
 	int i = 0, j = 0;
 	char table[41];
